refactor: User 조회용 서비스 인터페이스 추출
