+++ server/src/main/java/moment/user/application/UserQueryServiceImpl.java
@@ -1,0 +1,22 @@
+package moment.user.application;
+
+import lombok.RequiredArgsConstructor;
+import moment.global.exception.ErrorCode;
+import moment.global.exception.MomentException;
+import moment.user.domain.User;
+import moment.user.infrastructure.UserRepository;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class UserQueryServiceImpl {
+
+    private final UserRepository userRepository;
+
+    public User getUserById(Long id) {
+        return userRepository.findById(id)
+                .orElseThrow(() -> new MomentException(ErrorCode.USER_NOT_FOUND));
+    }
+}
